Convert src/components/HomeLink to typescript.
